requirements.txt
Flask==3.0.2
flask_cors==4.0.0  # need when gets cors errors
python-dotenv==1.0.1
pandas==2.2.2